Migrate shared code into a TestHelper module
